comment->recomment / edit&delete article, comment, recomment

+++ practice/session8/myproject/myapp/views.py
@@ -38,7 +38,6 @@ def login(request):
             error = '아이디 또는 비밀번호가 틀립니다'
             return render(request, 'registration/login.html', {'error': error})
 
-
     return render(request, 'registration/login.html')
 
 def logout(request):
@@ -47,62 +46,56 @@ def logout(request):
 
 @login_required(login_url="/registration/login/")
 def new(request):
-
     if request.method == 'POST':
         new_article = Article.objects.create(
             title = request.POST['title'],
             category = request.POST['category'],
             content = request.POST['content'],
-            author = request.user
+            author=request.user
         )
         return redirect('list')
     
     return render(request, 'new.html')
 
-def delete_article(request, article_id):
-    article = Article.objects.get(id=article_id)
-    if request.user == Article.objects.get(id=article_id).author:
-        Article.objects.filter(id=article_id).delete()
-        return redirect('list')
-    else:
-        error = "작성자만 삭제할 수 있습니다!"
-        return render(request, 'detail.html', {'article': article, 'error': error})
-    
 def edit_article(request, article_id):
     article = Article.objects.get(id=article_id)
     if request.method == 'POST':
-        if request.user == Article.objects.get(id=article_id).author:
-            Article.objects.filter(id=article_id).update(
-                title = request.POST['title'],
-                category = request.POST['category'],
-                content = request.POST['content'],
-            )
-            return render(request, 'detail.html', {'article': article})
-        else:
-            return render(request, 'detail.html', {'article': article})
+        Article.objects.filter(pk = article_id).update(
+            title = request.POST['title'],
+            category = request.POST['category'],
+            content = request.POST['content'],
+            author=request.user
+        )
+        return redirect('detail', article_id)
     return render(request, 'edit_article.html', {'article': article})
-    
+
+def delete_article(request, article_id):
+    article = Article.objects.get(id=article_id)
+    if article.author == request.user:
+        article.delete()
+        return redirect('list')
+    else:
+        error = "자신의 게시글만 삭제할 수 있습니다."
+        return redirect('detail_with_error', article_id=article_id, error=error)
+
 def list(request):
     articles = Article.objects.all()
-
-    return render(request, 'list.html', {'articles': articles, })
+    return render(request, 'list.html', {'articles': articles})
 
 @login_required(login_url="/registration/login/")
-def detail(request, article_id):
+def detail(request, article_id, error=None):
     article = Article.objects.get(id = article_id)
     comments = Comment.objects.filter(article = article)
-    recomments = Recomment.objects.exclude(parent_comment = None).filter(article = article)
+    recomments = Recomment.objects.all()
 
-    return render(request, 'detail.html', {'article': article, 'comments': comments, 'recomments': recomments})
+    if error:
+        return render(request, 'detail.html', {'article': article, 'comments': comments, 'recomments': recomments, 'error': error})
+    else:
+        return render(request, 'detail.html', {'article': article, 'comments': comments, 'recomments': recomments})
 
 def category(request, category):
     articles = Article.objects.filter(category = category)
     return render(request, 'category.html', {'articles': articles, 'category': category})
-
-def delete_comment(request, article_id, comment_id):
-    comment = Comment.objects.get(id=comment_id)
-    comment.delete()
-    return redirect('detail', article_id)
 
 def create_comment(request, article_id):
     if request.method == 'POST':
@@ -111,26 +104,40 @@ def create_comment(request, article_id):
         Comment.objects.create(
             article = article,
             content = content,
-            author = request.user
+            author=request.user
         )
         return redirect('detail', article_id)
     return render(request, 'detail.html', {'article': article})
 
-def create_recomment(request, article_id, comment_id):
+def delete_comment(request, article_id, comment_id):
     article = Article.objects.get(id=article_id)
+    if article.author == request.user:
+        comment = Comment.objects.get(pk=comment_id)
+        comment.delete()
+        return redirect('detail', article_id)
+    else:
+        error = "자신의 댓글만 삭제할 수 있습니다."
+        return redirect('detail_with_error', article_id=article_id, error=error)
+
+def create_recomment(request, article_id, comment_id):
+    article = Article.objects.get(pk=article_id)
     if request.method == 'POST':
         comment = Comment.objects.get(id = comment_id)
         content = request.POST['re_content']
         Recomment.objects.create(
-            article = article,
             content = content,
             parent_comment = comment,
-            author = request.user
+            author=request.user
         )
         return redirect('detail', article_id)
     return render(request, 'detail.html', {'article': article})
 
 def delete_recomment(request, article_id, recomment_id):
-    comment = Comment.objects.get(id=recomment_id)
-    comment.delete()
-    return redirect('detail', article_id)
+    article = Article.objects.get(id=article_id)
+    if article.author == request.user:
+        recomment = Recomment.objects.get(id = recomment_id)
+        recomment.delete()
+        return redirect('detail', article_id)
+    else:
+        error = "자신의 대댓글만 삭제할 수 있습니다."
+        return redirect('detail_with_error', article_id=article_id, error=error)
